add simpleSearchConditions.ts
