+++ AGENTS.md
@@ -98,6 +98,7 @@ app/frontend/
 - 検索条件のURL反映や初期復元は `app/frontend/src/store/searchManager.ts` と `initializeApp.ts` を確認する。
 - API通信は `app/frontend/src/api/fetchData.ts` など既存API層に合わせる。
 - StoreはAPIを直接呼ばない。仮想スクロールで未取得ページが必要な場合も、コンポーネントから `searchManager.requestNextPage()` を呼び、fetch の起動は `searchManager` / `api/fetchData.ts` 側へ集約する。
+- Simple Search条件のdefault差分抽出は `store/simpleSearchConditions.ts` に置く。`api/fetchData.ts` から `searchManager.ts` を import すると循環依存になるため避ける。
 - コンポーネントやViewから直接URLやfetchの仕様を増やす前に、既存のStore/API層へ寄せられるか確認する。
 - 戻る/進む、URL貼り付け、モード切り替えでは、Store更新と検索実行が重複しやすい。変更時は初期表示・タブ切替・履歴操作を分けて考える。
 - searchMode は内部リセットと検索開始副作用の順序が重要なため、通常操作では `storeManager.setSearchMode(mode)` を使う。
